feat(): encapsulation
make helper methods private

[ #121813963]
