+++ lib/pesuz/routing/mapper.rb
@@ -8,6 +8,7 @@ module Pesuz
       def map_to_route(request)
         @request = request
         path = request.path_info
+
         method = request.request_method.downcase.to_sym
         result = @endpoints[method].detect do |endpoint|
           match_path_with_pattern path, endpoint
